+++ src/App.tsx
@@ -4,7 +4,6 @@ import {Navigate, Route, Routes} from "react-router-dom";
 import "./assets/css/main.scss";
 import Layout from "./pages/layout/Layout";
 import SecretsList from "./pages/secrets-list/SecretsList";
-import SecretCreate from "./pages/secrets-list/SecretCreate";
 import Secrets from "./pages/secrets-list/Secrets";
 import {useAppDispatch, useAppSelector} from "./app/hooks";
 import SingUp from "./pages/authentication/SingUp";
@@ -54,7 +53,6 @@ function App() {
                         <Route element={<Layout/>}>
                             <Route path="secrets-list" element={<SecretsList/>}>
                                 <Route path="secrets" element={<Secrets/>}/>
-                                <Route path="secret-create" element={<SecretCreate/>}/>
                             </Route>
                         </Route>
                         <Route path="*" element={<Navigate replace to="/secrets-list"/>}/>
